requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
packaging==23.2
rich==13.7.0
pyvis==0.3.2
networkx==3.2
